requirements.txt
PyQt5==5.15.11
selenium==4.40.0
psycopg2-binary==2.9.11
webdriver-manager==4.0.2
pytest==8.4.2
